requirements.txt
asgiref==3.8.1
beautifulsoup4==4.12.3
crispy-bootstrap5==2024.2
dj-database-url==2.2.0
Django==4.2.13
django-bootstrap-form==3.4
django-bootstrap-v5==1.0.11
django-cors-headers==4.3.1
django-crispy-forms==2.1
django-environ==0.11.2
django-filter==24.3
djangorestframework==3.15.1
djangorestframework-simplejwt==5.3.1
docx==0.2.4
et-xmlfile==1.1.0
gunicorn==23.0.0
lxml==5.3.0
numpy==1.26.4
openpyxl==3.1.2
packaging==24.1
pandas==2.2.2
pillow==11.0.0
psycopg2-binary==2.9.9
PyJWT==2.9.0
python-dateutil==2.9.0.post0
python-docx==1.1.2
python-dotenv==1.0.1
pytz==2024.1
six==1.16.0
soupsieve==2.5
sqlparse==0.4.4
typing_extensions==4.12.2
tzdata==2024.1
whitenoise==6.7.0
